requirements.txt
Flask
Pillow
numpy
keras
